requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-crispy-forms==1.14.0
djangorestframework==3.13.1
future==0.18.2
pytz==2022.1
serializers==0.2.4
sqlparse==0.4.2
tzdata==2022.1
